Remove addEvent() from court
- Event is the owning side of the relationship with a NOT NULL foreign key of
a Court. Therefore, prior to persisting a new Event, `setCourt()` needs to be
invoked on it. Thus, Court's `addEvent()` is useless when creating an event.
- The after-creation reassignment of an Event to a different Court is a valid
use-case, however directly updating an Event using `setCourt()` -- rather than
removing it from one Court and adding it to another -- seems more
straightforward and sufficient. So `addEvent()` seems useless even for updates.